tasks 1-4
